5.0发布

+++ src/com/redmoon/oa/android/sales/ProductDetailAction.java
@@ -78,7 +78,7 @@ public class ProductDetailAction {
 							.next();
 					String name = am.getName();
 					String downUrl = "/public/visual/visual_getfile.jsp";
-					int attId = am.getId();
+					long attId = am.getId();
 					fileObj.put("name", name);
 					fileObj.put("downloadUrl", downUrl);
 					fileObj.put("id", attId);
